Address some changes requested in PR review
