+++ web/app/js/components/TapEventTable.jsx
@@ -15,6 +15,7 @@ import Typography from '@material-ui/core/Typography';
 import _get from 'lodash/get';
 import _isEmpty from 'lodash/isEmpty';
 import _isNull from 'lodash/isNull';
+import { headersDisplay } from './TapEventHeadersTable.jsx';
 import { withContext } from './util/AppContext.jsx';
 import { withStyles } from '@material-ui/core/styles';
 
@@ -124,65 +125,6 @@ const tapColumns = (resourceType, ResourceLink) => {
 
 const formatTapLatency = str => {
   return formatLatencySec(str.replace("s", ""));
-};
-
-const headersStyles = {
-  headerName: {
-    fontSize: "12px",
-    marginTop: "5px",
-  },
-};
-
-const HeadersContentBase = ({headers, classes}) => {
-  return (
-    <React.Fragment>
-      {headers.map(header => {
-        return (
-          <React.Fragment key={`${header.name}_${header.valueStr}`}>
-            <Typography
-              className={classes.headerName}
-              variant="inherit"
-              color="textPrimary">
-              {header.name}
-            </Typography>
-            <Typography
-              variant="inherit"
-              color="textSecondary">
-              {header.valueStr}
-            </Typography>
-          </React.Fragment>
-        );
-      })}
-    </React.Fragment>
-  );
-};
-
-HeadersContentBase.propTypes = {
-  classes: PropTypes.shape({}).isRequired,
-  headers: PropTypes.arrayOf(PropTypes.shape({
-    name: PropTypes.string.isRequired,
-    valueStr: PropTypes.string.isRequired,
-  })).isRequired,
-};
-
-HeadersContentBase.defaultProps = {
-  headers: [],
-};
-
-const HeadersContentDisplay = withStyles(headersStyles)(HeadersContentBase);
-
-const headersDisplay = (title, value) => {
-  if (!value) {
-    return null;
-  }
-
-  return (
-    <ListItem disableGutters>
-      <ListItemText
-        primary={title}
-        secondary={"headers" in value ? <HeadersContentDisplay headers={value.headers} /> : "-"} />
-    </ListItem>
-  );
 };
 
 const itemDisplay = (title, value) => {
